+++ src/main/java/nextstep/ladder/BooleanGenerator.java
